# docs/T033_openapi.yaml
openapi: "3.0.3"
info:
  title: GarudaX Platform API
  description: |
    REST API for the AI Powered Commodity Exchange of Mongolia.
    This gateway translates REST/JSON requests to gRPC calls on backend exchange services.
  version: "1.0.0"
  contact:
    name: GarudaX Platform Engineering
  license:
    name: Proprietary
servers:
  - url: https://api.garudax.mn/api/v1
    description: Production
  - url: https://staging-api.garudax.mn/api/v1
    description: Staging
...
security:
  - BearerAuth: []
paths:
...
  /margin:
    get:
      tags: [Margin]
      summary: Get portfolio margin for authenticated participant
      operationId: getPortfolioMargin
      parameters:
        - name: participant_id
          in: query
          schema:
            type: string
            format: uuid
            description: Defaults to authenticated user's participant_id
      responses:
        "200":
          description: Portfolio margin details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortfolioMargin"
        "404":
          $ref: "#/components/responses/NotFound"
  /margin/calculate:
    post:
      tags: [Margin]
      summary: Trigger margin calculation
      operationId: calculateMargin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                participant_id:
                  type: string
                  format: uuid
              required: [participant_id]
      responses:
        "200":
          description: Margin calculation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortfolioMargin"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  reason:
                    type: string
            request_id:
              type: string
          required: [code, message]
...
    MarginRequirement:
      type: object
      properties:
        participant_id:
          type: string
          format: uuid
        instrument_id:
          type: string
          format: uuid
        net_quantity:
          type: integer
        scan_risk:
          type: string
        inter_month:
          type: string
        delivery_month:
          type: string
        short_option:
          type: string
        initial_margin:
          type: string
        mark_to_market:
          type: string
        total_required:
          type: string
        calculated_at:
          type: string
          format: date-time
    PortfolioMargin:
      type: object
      properties:
        participant_id:
          type: string
          format: uuid
        requirements:
          type: array
          items:
            $ref: "#/components/schemas/MarginRequirement"
        total_initial:
          type: string
        total_mtm:
          type: string
        total_required:
          type: string
        collateral_on_hand:
          type: string
        excess_deficit:
          type: string
        calculated_at:
          type: string
          format: date-time
...